+++ modules/core/src/main/scala/gem/syntax/package.scala
@@ -1,0 +1,8 @@
+// Copyright (c) 2016-2017 Association of Universities for Research in Astronomy, Inc. (AURA)
+// For license information see LICENSE or https://opensource.org/licenses/BSD-3-Clause
+
+package gem
+
+package object syntax {
+  object all extends ToParserOps
+}
